:seedling: A new dawn for `6utt3r`...
I'm planning on redoing this project using C and SDL2, as intended
originally. This is because I kinda messed up with the loading times and
the Python implementation was supposed to be temporary and for
presentation's purposes. Claiming that the Python Code was originally
mine is a great sin and I should be punished for that.

Issue 1