:sparkles: Implemented Surface modifier!
This implementation features a means of drawing pixels onto the screen
as well as a few additions to the header file. When pressing LMB and
RMB, a red and blue pixel is activated, respectively. This took some
time to implement, but I'm pretty stoked we had something working.
Cheers.

TODO: Implement a Render and Draw function.