:sparkles: [rectangle] Now draws a rectangle to screen!
We did it guys! We drew a rectangle to the center of the screen! This
will form the basis of building the map and the engine itself! Finally
dude!
Unfortunately as it's now, we have a SIGSEGV when we use the mouse and
try to place pixels to screen. So...this will be fixed in the following
commits. Stay tuned for more!

TODO: Make Rectangle textured!