feat: (refactor #52) Opted for simplicity and no emojis
Review #1:
So it turns out that emojis are pretty unprofessional and made my
Commit messages weird and misunderstood. I need to ensure that my
Commits follow SemVer and Conventional Commits style for the sake of the
people who'll be reviewing my work. So this is goodbye to emojis and the
Ikiara Commit Style Guide!

Review #52:
The SIGSEGV and Linking Issues were caused by the "Complex" & "Uninformed"
Abstraction that was originally implemented. I hope that this commit
will be a reflection of this simplified C code that should work without
error. Cheers!

TODO: - Render a Textured Rectangle onto the Window.
      - Reorganize the code and make it SOLID & DRY.